Replace `letInEvents` with `supplyOn` and `supplyAfter`
