+++ src/processors/index.ts
@@ -3,13 +3,14 @@ export * from './dig-after';
 export * from './dig-on';
 export * from './event-mapper';
 export * from './filter-on';
-export * from './let-in-events';
 export * from './map-events';
 export * from './once-after';
 export * from './once-on';
 export * from './resolve-events';
 export * from './resolve-events-in-order';
 export * from './share-events';
+export * from './supply-after';
+export * from './supply-on';
 export * from './translate-after';
 export * from './translate-on';
 export * from './value-events';
