+++ src/processors/once-after.spec.ts
@@ -1,7 +1,7 @@
 import { neverSupply, Supply } from '@proc7ts/primitives';
 import { AfterEvent, afterEventBy } from '../after-event';
 import { EventNotifier, EventReceiver } from '../base';
-import { onceOn } from './once-on';
+import { onceAfter } from './once-after';
 
 describe('onceAfter', () => {
 
@@ -25,35 +25,35 @@ describe('onceAfter', () => {
   });
 
   it('registers event receiver', () => {
-    afterEvent.do(onceOn)(mockReceiver);
+    afterEvent.do(onceAfter)(mockReceiver);
     expect(mockRegister).toHaveBeenCalled();
   });
   it('sends initial event', () => {
-    afterEvent.do(onceOn)(mockReceiver);
+    afterEvent.do(onceAfter)(mockReceiver);
     expect(mockReceiver).toHaveBeenCalledWith('init');
   });
   it('cuts off supply after event received', () => {
 
-    const returnedSupply = afterEvent.do(onceOn)(mockReceiver);
+    const returnedSupply = afterEvent.do(onceAfter)(mockReceiver);
 
     expect(mockRegister).toHaveBeenCalled();
     expect(returnedSupply.isOff).toBe(true);
     expect(supply.isOff).toBe(true);
   });
   it('unregisters notified event receiver', () => {
-    afterEvent.do(onceOn)(mockReceiver);
+    afterEvent.do(onceAfter)(mockReceiver);
     expect(whenOff).toHaveBeenCalled();
   });
   it('never sends events if their supply is initially cut off', () => {
     supply = neverSupply();
-    afterEvent.do(onceOn)({
+    afterEvent.do(onceAfter)({
       supply,
       receive: (_context, ...event: [string]) => mockReceiver(...event),
     });
     expect(mockReceiver).not.toHaveBeenCalled();
   });
   it('sends only one event', () => {
-    afterEvent.do(onceOn)(mockReceiver);
+    afterEvent.do(onceAfter)(mockReceiver);
     emitter.send('event1');
     emitter.send('event2');
     expect(mockReceiver).toHaveBeenCalledTimes(1);
