+++ src/processors/resolve-events-in-order.ts
@@ -5,9 +5,9 @@
 import { Supply } from '@proc7ts/primitives';
 import { sendEventsTo } from '../base';
 import { OnEvent, onEventBy } from '../on-event';
-import { letInEvents } from './let-in-events';
 import { mapEvents } from './map-events';
 import { resolveEvents } from './resolve-events';
+import { supplyOn } from './supply-on';
 
 /**
  * A processor that asynchronously resolves incoming events and sends them in the order they are received.
@@ -36,7 +36,7 @@ export function resolveEventsInOrder<TEvent>(
     const sourceSupply = new Supply();
     let numInProcess = 0;
     const source = from.do(
-        letInEvents(supply, sourceSupply),
+        supplyOn(supply, sourceSupply),
         mapEvents(event => {
           ++numInProcess;
           return event;
