+++ src/processors/supply-on.ts
@@ -4,11 +4,11 @@
  */
 import { Supply, SupplyPeer } from '@proc7ts/primitives';
 import { eventLetIn } from '../impl';
-import { OnEvent } from '../on-event';
-import { EventMapper } from './event-mapper';
+import { OnEvent, onEventBy } from '../on-event';
 
 /**
- * Creates an event processor that passes incoming events until the `required` supply is cut off.
+ * Creates an event processor that passes events incoming from {@link OnEvent} sender until the `required` supply is
+ * cut off.
  *
  * The outgoing events supply will be cut off once incoming event supply does, unless a second supply passed in.
  * In the latter case that supply will be cut off instead.
@@ -20,11 +20,9 @@ import { EventMapper } from './event-mapper';
  *
  * @returns New event mapper.
  */
-export function letInEvents<TEvent extends any[]>(
+export function supplyOn<TEvent extends any[]>(
     required: SupplyPeer,
     dependentSupply?: Supply,
-): EventMapper<TEvent> {
-  return (
-      (input: OnEvent<TEvent>) => input.by(eventLetIn(input, required, dependentSupply))
-  ) as EventMapper<TEvent>;
+): (this: void, input: OnEvent<TEvent>) => OnEvent<TEvent> {
+  return (input: OnEvent<TEvent>) => onEventBy(eventLetIn(input, required, dependentSupply));
 }
